Central Brisbane surface clay content

A rasterStack containing modelled estimated percent clay content for central Brisbane, in South East Queensland.

bne_surface_clay

Format

An object of class RasterStack of dimension 121 x 145 x 3.

Details

The dataset was retrieved from the National Soil Attributes Clay WCS on 2019/07/07 using the demonstration code in get_soils_data.

The dataset has three named layers. The first is the estimated value, the second is the 5% confidence limit, and the third is the 95% confidence limit.

The dataset is in WGS84 (EPSG:4326) and has a resolution of 3 arc seconds, which is approximately 80x90m when projected into EPSG:28355 or EPSG:3577.

Note that some off-shore areas have a value of 0 rather than NA. A coastline masking layer will be required to safely remove these values.